+++ apps-intern/ungdomsytelse-veileder/src/types/Features.ts
@@ -1,7 +1,6 @@
 import { appEnv } from '../utils/appEnv';
 
 export const Features = {
-    endreSluttdato: appEnv.SIF_PUBLIC_FEATURE_ENDRE_SLUTTDATO === 'on',
-    sjekkliste: appEnv.SIF_PUBLIC_FEATURE_SJEKKLISTE === 'on',
+    utvidePeriode: appEnv.SIF_PUBLIC_FEATURE_UTVIDE_PERIODE === 'on',
     slettAktivDeltakelse: appEnv.SIF_PUBLIC_FEATURE_SLETT_AKTIV_DELTAKELSE === 'on',
 };