+++ packages/ung-brukerdialog-api/src/ung-brukerdialog-api/client/types.gen.ts
@@ -10,12 +10,6 @@ export type ArbeidOgFrilansRegisterInntektDto = {
     inntekt: number;
 };
 
-export type BekreftBostedOppgavetypeDataDto = {
-    erBosattITrondheim: boolean;
-    fom: string;
-    tom: string;
-};
-
 export type BrukerdialogOppgaveDto = {
     frist?: string;
     løstDato?: string;
@@ -116,10 +110,6 @@ export enum OppgaveType {
      * SØK_YTELSE
      */
     SØK_YTELSE = 'SØK_YTELSE',
-    /**
-     * BEKREFT_BOSTED
-     */
-    BEKREFT_BOSTED = 'BEKREFT_BOSTED',
 }
 
 export enum OppgaveYtelsetype {
@@ -134,9 +124,6 @@ export enum OppgaveYtelsetype {
 }
 
 export type OppgavetypeDataDto = (
-    | ({
-          type: 'BOSTED';
-      } & BekreftBostedOppgavetypeDataDto)
     | ({
           type: 'ENDRET_PERIODE';
       } & EndretPeriodeDataDto)