+++ packages/ung-deltakelse-opplyser-api-veileder/src/veileder/client/zod.gen.ts
@@ -21,7 +21,9 @@ export const zDeltakelseDto = z.object({
     erSlettet: z.boolean(),
     fraOgMed: z.iso.date(),
     harOpphørsvedtak: z.boolean(),
+    harUtvidetKvote: z.boolean(),
     id: z.uuid().nullish(),
+    kvoteMaksDato: z.iso.date(),
     søktTidspunkt: z.iso.datetime({ local: true }).nullish(),
     tilOgMed: z.iso.date().nullish(),
 });
@@ -39,6 +41,7 @@ export const zEndringstype = z.enum([
     'ENDRET_SLUTTDATO',
     'DELTAKER_HAR_SØKT_YTELSE',
     'DELTAKELSE_FJERNET',
+    'UTVIDET_KVOTE',
     'UKJENT',
 ]);
 
@@ -138,6 +141,15 @@ export const zDeltakelseHistorikkPath = z.object({
  * OK
  */
 export const zDeltakelseHistorikkResponse = z.array(zDeltakelseHistorikkDto);
+
+export const zUtvidKvotePath = z.object({
+    deltakelseId: z.uuid(),
+});
+
+/**
+ * OK
+ */
+export const zUtvidKvoteResponse = zDeltakelseDto;
 
 export const zMeldInnDeltakerBody = zDeltakelseInnmeldingDto;