+++ apps/aktivitetspenger-soknad/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @navikt/aktivitetspenger-soknad
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.3.19
+    - @navikt/k9-brukerdialog-prosessering-api@0.3.16
+    - @navikt/k9-sak-innsyn-api@0.2.18
+    - @sif/api@0.0.5
+    - @sif/soknad-forms@0.0.5
+
+## 0.0.7
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @sif/soknad-forms@0.0.4
+    - @sif/soknad@0.0.4
+    - @sif/api@0.0.4
+    - @sif/rhf@0.0.3
+
 ## 0.0.6
 
 ### Patch Changes