+++ packages/sif-soknad-forms/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sif/soknad-forms
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+    - @sif/api@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @sif/api@0.0.4
+    - @sif/rhf@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes